code generator: removed code treating data members as a vector if all the data members are of the same type (for example, in stamp, sec and nanosec were returned as a vector) -- for sequences, removed suffix "Vector" in accessor methods -- promoted ROS types to simpler C++ types so VTK python wrappers work.

+++ MRML/CodeGeneration/configForCodegen.py
@@ -1,9 +1,11 @@
+# mapping from ROS to vtk/C++ types with upcast to Python wrappers are
+# generated
 static_type_mapping = {
     'bool': 'bool',
-    'uint32': 'uint32_t',
-    'int32': 'int32_t',
-    'int' : 'int32_t',
-    'float': 'float',
+    'uint32': 'int',
+    'int32': 'int',
+    'int' : 'int',
+    'float': 'double',
     'double': 'double',
     'string': 'std::string',
     'str': 'std::string'
